RPiServer/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
requests==2.21.0
RPi.GPIO==0.7.0
urllib3==1.24.3
Werkzeug==0.16.0
